Moved gameMap interface to logiclayer and used it
