+++ src/main/Coordinator.java
@@ -1,18 +1,18 @@
 package main;
 import logicLayer.GameLogic;
-import logicLayer.GameMap;
+import logicLayer.GameMapImpl;
 import logicLayer.Position;
 import inputLayer.TextInputReciever;
 import output.TextOutputPrinter;
 
 public class Coordinator {
 	
-	private GameMap gameMap;
+	private GameMapImpl gameMap;
 	private TextInputReciever textInput = new TextInputReciever();
 	private TextOutputPrinter textOutput = new TextOutputPrinter();
 	private GameLogic logic;
 	
-	public Coordinator(GameMap gameMap) {
+	public Coordinator(GameMapImpl gameMap) {
 		this.gameMap = gameMap;
 		this.logic = new GameLogic(gameMap);
 	}
