--- src/output/GameMap.java
@@ -1,5 +1,0 @@
-package output;
-
-public interface GameMap {
-	void incrementTurn();
-}
